# Conflicts:
#	VerticalWarehouses/Machine Automation Service/Ferretto.VW.MAS.AutomationService/Configuration/vertimag-configuration.Simulation.json
#	VerticalWarehouses/Machine Automation Service/Ferretto.VW.MAS.AutomationService/Controllers/MachineSetup/BeltBurnishingProcedureController.cs
#	VerticalWarehouses/Machine Automation Service/Ferretto.VW.MAS.AutomationService/Controllers/MachineSetup/DepositAndPickupProcedureController.cs
#	VerticalWarehouses/Machine Automation Service/Ferretto.VW.MAS.AutomationService/Controllers/ShuttersController.cs
#	VerticalWarehouses/Machine Automation Service/Ferretto.VW.MAS.DeviceManager/Providers/ElevatorProvider.cs

+++ VerticalWarehouses/Machine Automation Service/Ferretto.VW.MAS.DeviceManager/Positioning/PositioningExecutingState.cs
@@ -211,7 +211,7 @@ namespace Ferretto.VW.MAS.DeviceManager.Positioning
                 FieldMessageActor.InverterDriver,
                 FieldMessageActor.FiniteStateMachines,
                 FieldMessageType.InverterSetTimer,
-                (byte)inverterIndex));
+                inverterIndex));
         }
 
         public override void Stop(StopRequestReason reason)
